vmm: move all error related logic to a file
All this commit does is moving error related logic
to a new file under vmm/src/error.rs.
vmm/src/lib.rs is now around 3200 lines (4000 before).
Improved coverage for this newly added file.
